+++ scratch.js
@@ -3,26 +3,12 @@ const nlp = require('./src/index')
 // nlp.extend(require('./plugins/sentences/src'))
 // nlp.extend(require('./plugins/numbers/src'))
 
-// let doc = nlp(`i walked to the moon when it was shining`)
-// doc.verbs().forEach(d => {
-//   d.matchOne('walked').replaceWith('sat')
-// })
-// doc.sentences().toPastTense()
-// doc.debug()
+// nlp(`wayne's world, party time`)
+//   .match('#Noun+? wayne')
+//   .debug()
 
-// let doc = nlp.tokenize(`between june 5th and june 7th`)
-// doc.match('between [#Date+] and').debug()
-// doc.match('between [.*] and').debug()
+// console.log('aaaab'.match(/a+?[a|b]/))
 
-// let doc = nlp('falls over')
-// let doc = nlp('i paid $5.20 for the thing')
-// let doc = nlp('i got 1 peso and £30.').debug()
-// doc.money().debug()
-
-// nlp.verbose('tagger')
-
-// nlp('His is green.').debug()
-
-// nlp('her glue').debug()
-// nlp('let Toronto Ontario glue it').debug()
-nlp(`i think he's cool`)
+nlp('Toronto Ontario foo')
+  .match('#Place+ .')
+  .debug()